tests/submission/Microsoft/canvas/canvas_colorsAndStyles_createPattern_001.htm

<!DOCTYPE HTML>
<html>
    <head>
        <title>HTML5 Canvas Test:  createPattern() SYNTAX_ERR exception</title>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com" />
        <link rel="help" href="http://www.w3.org/TR/2dcontext/#dom-context-2d-createpattern" />
        <meta name="assert" content="If the repetition argument in createPattern(image, repetition) is unrecognized, SYNTAX_ERR exception must be raised."/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
        <script type="text/javascript">
            setup({explicit_done: true});
            function runTest() 
            {
              test(function()
              {
                var canvas = document.getElementById("canvas1");
                var ctx = canvas.getContext("2d");

                var x;
                var img = document.getElementById("imgBlackRect");

                // SYNTAX_ERR exception should be thrown if repetition argument is unrecognized.
                assert_throws("SYNTAX_ERR", function () { ctx.createPattern(img, x); } );
              }, "unrecognized repetition argument");
              done();
            }
        </script>
    </head>
    <body onload="runTest()">
        <p>Description:  If the repetition argument in createPattern(image, repetition) is unrecognized, SYNTAX_ERR exception must be raised.</p>
        <div id="log"></div>
        <canvas id="canvas1" width="300" height="150">Browser does not support HTML5 Canvas.</canvas>
        <img id="imgBlackRect" style="display:none" src="/support/images/black-rectangle.png">
    </body>
</html>
